\name{mibrrW}
\alias{mibrrW}
\title{Print Warranty Statement for \pkg{mibrr}}
\description{
  Print the sections of the GPL-3 that describe the warranty (or complete lack
  thereof) for \pkg{mibrr}. 
}
\usage{
mibrrW()
}
\value{
  Text giving the warranty-specific sections of the GPL-3.
}
\author{
  Kyle M. Lang
}
\examples{
mibrrW()
}